improved profiling driver, apparently radix has issues with really big arrays (2^18 or greater)

+++ brick.hpp
@@ -1,5 +1,5 @@
-void brickSort(int arr[], int n);
-void quickSort(int arr[], int n);
-void mergeSort(int arr[], int n);
-void radixSort(int arr[], int n);
-void bitonicSort(int arr[], int n);
+void brickSort(int64_t arr[], int64_t n);
+void quickSort(int64_t arr[], int64_t n);
+void mergeSort(int64_t arr[], int64_t n);
+void radixSort(int64_t arr[], int64_t n);
+void bitonicSort(int64_t arr[], int64_t n);
